+++ microservices/review-service/src/main/java/com/github/valiclud/old_languages/core/review/services/ReviewServiceImpl.java
@@ -22,7 +22,7 @@ public class ReviewServiceImpl implements ReviewService {
   }
 
   @Override
-  public List<Review> getReviews(Long productId) {
+  public List<Review> getReviews(int productId) {
 
     if (productId < 1) {
       throw new InvalidInputException("Invalid productId: " + productId);
